feat: rename to `core`
