--- src/browser.ts
@@ -1,5 +1,0 @@
-export * from "./core";
-export * from "./types";
-export * from "./format";
-export * from "./error";
-export * from "./builders";
