+++ src/index.ts
@@ -1,6 +1,2 @@
-export * from "./code";
+export * from "./file";
 export * from "./core";
-export * from "./types";
-export * from "./format";
-export * from "./error";
-export * from "./builders";
