+++ test/object.test.ts
@@ -293,7 +293,6 @@ export default {
     const urlPattern = mod.exports.default.urlPattern;
     expect(typeof urlPattern).toBe("object");
     expect(urlPattern).toBeInstanceOf(RegExp);
-    // eslint-disable-next-line unicorn/prefer-string-raw
     expect(urlPattern.source).toBe("\\/api\\/pwa\\/.*");
 
     mod.exports.default.urlPattern = /\/api\/pwa\/v1\/.*/;
