+++ test/utils.test.ts
@@ -11,7 +11,6 @@ describe("literalToAst", () => {
     expect(run(1)).toMatchInlineSnapshot('"1"');
     expect(run(true)).toMatchInlineSnapshot('"true"');
     expect(run(undefined)).toMatchInlineSnapshot('"undefined"');
-    // eslint-disable-next-line unicorn/no-null
     expect(run(null)).toMatchInlineSnapshot('"null"');
     expect(run([undefined, 1, { foo: "bar" }])).toMatchInlineSnapshot(`
       "[undefined, 1, {
